Task: Switch to curl + HTTP caching + streamlined API
HTTP caching isn't very useful in prod, but it certainly does simplify
development and debugging.

The streamlined API allows for URL-fetching code to throw error objects
with intent (fatal error / expected error) and with the HTTP response as
context, which simplifies site checking code. Main advantage is that
this also makes it easier to extract and store response information in
the database, but I haven't implemented that yet.

Shelling out to curl is not ideal, but I've had some reservations about
LWP::UserAgent and curl is a fine alternative anyway.